Move DNS stuff to single new file rsdnsutils.cc.
